requirements.txt
asgiref==3.10.0
diff-match-patch==20241021
Django==5.0.7
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-cors-headers==4.4.0
django-import-export==4.3.10
django-js-asset==3.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==10.4.0
PyJWT==2.10.1
sqlparse==0.5.3
tablib==3.9.0
tzdata==2025.2
